# applications/rest-api/openapi.yaml
openapi: 3.0.0
info:
  title: Sample REST API
  version: 0.0.1
paths:
  /posts:
    get:
      summary: Lists posts
      description: Lists all posts subject to the supplied filters
      parameters:
        - name: created_after
          in: query
          description: Restricts the list of posts to those created after (inclusive) this datetime.
          schema:
            type: string
            format: datetime
        - name: created_before
          in: query
          description: Restricts the list of posts to those created before (exclusive) this datetime.
          schema:
            type: string
            format: datetime
        # TODO add pagination fields
      responses:
        200:
          description: A successful list of posts that met the supplied filters
          content:
            application/json:
              schema:
                type: object
                required:
                  - posts
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostResponse'
        500:
          $ref: '#/components/responses/InternalError'
        # TODO add error codes
    post:
      summary: Creates a post
      description: Creates a post
      requestBody:
        description: Details of the post to create
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        201:
          description: The successfully created post
          content:
            application/json:
              schema:
                type: object
                required:
                  - post
                properties:
                  post:
                    $ref: '#/components/schemas/PostResponse'
        500:
          $ref: '#/components/responses/InternalError'
        # TODO add error codes
  # TODO add get single, update and delete endpoints
components:
  schemas:
    PostResponse:
      type: object
      required:
        - id
        - title
        - content
        - created_at
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: datetime
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
  responses:
    InternalError:
      description: An internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'